feat: migrate project to TypeScript and restructure codebase
